requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
attrs==24.3.0
certifi==2024.12.14
frozenlist==1.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.18.3
